Improvements to the category screen composable
