+++ app/src/main/java/com/example/shopapp/composable/BottomBar.kt
@@ -24,7 +24,7 @@ fun BottomBar() {
         modifier = Modifier
             .fillMaxWidth()
             .background(MaterialTheme.colors.background)
-            .padding(vertical = 5.dp),
+            .padding(vertical = 10.dp),
         verticalAlignment = Alignment.CenterVertically,
         horizontalArrangement = Arrangement.SpaceAround
     ) {
